refactor(): move  QuestDetailsModal to quest card component
