+++ src/modules/canvas/components/quest-card-component/QuestDetailsModal.tsx
